refactor: rename files; updated resolve path for webpack
