+++ src/app/requestListenerWithErrorHandling.ts
@@ -1,12 +1,12 @@
 import { RequestListener } from 'http';
 
-import { prepareUserRoutes } from '../components/users/user.router';
+import { prepareUserRoutes } from '../components/users/router';
 import Router from '../shared/server/router';
 import { AppError, NotFoundError } from '../shared/error/errorInstance';
 import { combineResponseWithError } from '../shared/server/utils/combineResponse';
 import { getRouteParts } from '../shared/server/utils/getRouteParts';
-import { API_URL_BASE } from '../shared/server/http.constants';
-import { IClientIncomingMessage } from '../shared/server/server.interface';
+import { API_URL_BASE } from '../shared/server/constants';
+import { IClientIncomingMessage } from '../shared/server/interface';
 
 export const requestListenerWithErrorHandling: RequestListener = async (request, response) => {
   if (request.url === undefined) {
