+++ src/components/users/storage.ts
@@ -1,8 +1,8 @@
 import { v4 as uuidv4 } from 'uuid';
 
-import { IUserResponse, IUserRequest } from './user.interface';
+import { IUserResponse, IUserRequest } from './interface';
 import { AppError, InternalError } from '../../shared/error/errorInstance';
-import { COMMON_ERROR_MESSAGE } from '../../shared/error/error.message';
+import { COMMON_ERROR_MESSAGE } from '../../shared/error/messages';
 import { IGenericRepository } from '../../shared/storage/interface';
 import { GenericRepository } from '../../shared/storage/repository';
