--- src/shared/server/router.interface.ts
@@ -1,12 +1,0 @@
-import { ServerResponse } from 'http';
-
-import { METHOD } from './http.constants';
-import { IClientIncomingMessage } from './server.interface';
-
-export interface IRouteDefinition {
-  basePath: string;
-  execute: (request: IClientIncomingMessage, response: ServerResponse) => Promise<void>;
-  method: METHOD;
-  pathParts?: string[];
-  shouldHaveBody: boolean;
-}
