--- src/shared/server/server.interface.ts
@@ -1,9 +1,0 @@
-import { IncomingMessage } from 'http';
-
-export interface IClientIncomingMessage extends IncomingMessage {
-  url: string;
-}
-
-export interface IWorkerToPortMap {
-  [workerId: number]: number;
-}
