+++ src/shared/server/utils/requestBody.ts
@@ -3,7 +3,6 @@ import { IncomingMessage } from 'http';
 export const getRequestBody = async <TBody>(request: IncomingMessage): Promise<TBody | null> => {
   const buffers: Uint8Array[] = [];
 
-  // TODO
   // eslint-disable-next-line no-restricted-syntax
   for await (const chunk of request) {
     buffers.push(chunk);
