+++ src/shared/validation/specific/age.validator.ts
@@ -1,4 +1,4 @@
-import { INCORRECT_AGE_ERROR_MESSAGE } from '../../error/error.message';
+import { INCORRECT_AGE_ERROR_MESSAGE } from '../../error/messages';
 import { ValidationError } from '../../error/errorInstance';
 
 export const validateAge = (age: number, minAge: number): void => {
