+++ src/shared/validation/specific/uuid.validator.ts
@@ -1,6 +1,6 @@
 import { validate as uuidValidate } from 'uuid';
 
-import { INVALID_UUID_ERROR_MESSAGE } from '../../error/error.message';
+import { INVALID_UUID_ERROR_MESSAGE } from '../../error/messages';
 import { ValidationError } from '../../error/errorInstance';
 
 export const validateUuid = (uuid: string): void => {
